Address pager row mapping review feedback
