+++ src/Pages.php
@@ -11,7 +11,6 @@ use Ray\AuraSqlModule\Pagerfanta\ExtendedPdoAdapter;
 use Ray\AuraSqlModule\Pagerfanta\Page;
 use Ray\MediaQuery\Exception\LogicException;
 
-use function array_map;
 use function is_array;
 
 /** @template T of class-string|mixed */
@@ -56,17 +55,14 @@ final class Pages implements PagesInterface
     public function offsetGet($pageIndex): Page|null
     {
         $page = $this->delegate->offsetGet($pageIndex);
-        if (! $page instanceof Page || $this->rowMapper === null || ! is_array($page->data)) {
+        $data = $page instanceof Page ? $page->data : null;
+        if (! $page instanceof Page || $this->rowMapper === null || ! is_array($data)) {
             return $page;
         }
 
         $rowMapper = $this->rowMapper;
-        $page->data = array_map(
-            static function (mixed $row) use ($rowMapper): mixed {
-                return is_array($row) ? $rowMapper($row) : $row;
-            },
-            $page->data,
-        );
+        $page = clone $page;
+        $page->data = PageRows::map($data, $rowMapper);
 
         return $page;
     }
